+++ app/lib/models/user.g.dart
@@ -7,7 +7,7 @@ part of 'user.dart';
 // **************************************************************************
 
 User _$UserFromJson(Map<String, dynamic> json) => User(
-  id: (json['id'] as num).toInt(),
+  id: json['id'] as String,
   email: json['email'] as String,
   username: json['username'] as String,
   avatarUrl: json['avatar_url'] as String?,